Move getLikedSongsAsPlaylist function to mixit.ts
