+++ lib/mixit-apps.ts
@@ -2,12 +2,14 @@ import { MixitInput, ShuffleOption } from "@/types/mixit";
 import { Playlist } from "@/types/spotify";
 import {
   createPlaylistAndPopulateWithTracks,
-  getCurrentUser,
-  getLikedSongsAsPlaylist,
   replaceTracksInPlaylist,
 } from "./spotify-query";
 import { Session } from "next-auth";
-import { fisherYatesShuffle, getRandomSongsFromPlaylist } from "./mixit";
+import {
+  fisherYatesShuffle,
+  getRandomSongsFromPlaylist,
+  getLikedSongsAsPlaylist,
+} from "./mixit";
 
 type useShufflerAppParams = {
   data: MixitInput;
@@ -23,8 +25,9 @@ export async function useShufflerApp({
   (SpotifyApi.CreatePlaylistResponse | undefined) | string
 > {
   let playlist;
+
   if (data.input.type === "liked-songs") {
-    playlist = await getLikedSongsAsPlaylist(session);
+    playlist = await getLikedSongsAsPlaylist(session, shuffleOptions);
   } else {
     playlist = data.input.playlist as Playlist;
   }
