+++ lib/spotify-query.ts
@@ -94,41 +94,6 @@ export async function getCurrentUser({
 
 export const MAX_REQUESTS_FOR_LIKED_SONGS = 10;
 
-export async function getLikedSongsAsPlaylist(
-  session: Session
-): Promise<LikedSongsPlaylist> {
-  if (!session) {
-    signIn("spotify");
-    return Promise.reject("No session");
-  }
-
-  const spotify = getSpotifyApi(session);
-
-  // Get the first song so something is requested
-  const likedSongResponse = await spotify.getMySavedTracks();
-  const likedSongs = likedSongResponse.body.items;
-
-  if (likedSongResponse.statusCode !== 200) {
-    Promise.reject("Error getting user");
-  }
-
-  const userId = (await spotifyApi.getMe()).body.id;
-  const totalLikedSongs = likedSongResponse.body.total;
-
-  const likedSongsPlaylist = {
-    name: "Liked Songs",
-    id: "liked-songs-id",
-    userId: userId,
-    tracks: {
-      href: "https://api.spotify.com/v1/me/tracks",
-      total: totalLikedSongs,
-      items: likedSongs,
-    },
-  } as LikedSongsPlaylist;
-
-  return likedSongsPlaylist;
-}
-
 export const MAX_PLAYLIST_LENGTH = 100;
 export const MAX_SONGS_PER_REQUEST = 50;
